Move each test case to its own package.
