+++ lens4j-processor/src/test/resources/cases/decapitalize_factory_name/AccountWithDeCapitalizeSpecificFactoryName.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.decapitalize_factory_name;
 
 import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
