+++ lens4j-processor/src/test/resources/cases/empty_lens_name/AccountWithEmptyLensName.java
@@ -1,23 +1,19 @@
-package util.examples;
+package cases.empty_lens_name;
 
 import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
 import dev.khbd.lens4j.core.annotations.Lens;
 import dev.khbd.lens4j.core.annotations.LensType;
 
-@GenLenses(
-        lenses = {
-                @Lens(path = "accountNumber"),
-                @Lens(path = "accountNumber", lensName = "ACCOUNT_NUMBER_READ_WRITE_LENS", type = LensType.READ_WRITE),
-                @Lens(path = "currency.code", lensName = "ACCOUNT_CURRENCY_CODE_READ_LENS"),
-                @Lens(path = "currency.id", lensName = "ACCOUNT_CURRENCY_CODE_READ_WRITE_LENS", type = LensType.READ_WRITE)
-        }
-)
-class PackagePrivateAccount {
+@GenLenses(lenses = {
+        @Lens(path = "currencyFieldForTest.code"),
+        @Lens(path = "currencyFieldForTest.id", type = LensType.READ_WRITE)
+})
+public class AccountWithEmptyLensName {
     private String id;
     private String accountNumber;
     private String bic;
-    private Currency currency;
+    private Currency currencyFieldForTest;
 
     public void setId(String id) {
         this.id = id;
@@ -43,11 +39,11 @@ class PackagePrivateAccount {
         return bic;
     }
 
-    public void setCurrency(Currency currency) {
-        this.currency = currency;
+    public void setCurrencyFieldForTest(Currency currencyFieldForTest) {
+        this.currencyFieldForTest = currencyFieldForTest;
     }
 
-    public Currency getCurrency() {
-        return currency;
+    public Currency getCurrencyFieldForTest() {
+        return currencyFieldForTest;
     }
 }
