+++ lens4j-processor/src/test/resources/cases/empty_lens_name/AccountWithEmptyLensNameLenses.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.empty_lens_name;
 
 import common.Currency;
 import dev.khbd.lens4j.core.Lenses;
