+++ lens4j-processor/src/test/resources/cases/inner_class/GenLensesOnInnerClass.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.inner_class;
 
 import dev.khbd.lens4j.core.annotations.GenLenses;
 import dev.khbd.lens4j.core.annotations.Lens;
