+++ lens4j-processor/src/test/resources/cases/multi_property/Payment.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.multi_property;
 
 import common.Currency;
 import common.Receiver;
