+++ lens4j-processor/src/test/resources/cases/multi_property/PaymentLenses.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.multi_property;
 
 import common.Bank;
 import common.Payer;
