+++ lens4j-processor/src/test/resources/cases/on_interface/LensOnInterface.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.on_interface;
 
 import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
