+++ lens4j-processor/src/test/resources/cases/package_private_class/PackagePrivateAccount.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.package_private_class;
 
 import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
@@ -6,13 +6,14 @@ import dev.khbd.lens4j.core.annotations.Lens;
 import dev.khbd.lens4j.core.annotations.LensType;
 
 @GenLenses(
-        factoryName = "SpecificFactoryName",
         lenses = {
+                @Lens(path = "accountNumber"),
+                @Lens(path = "accountNumber", lensName = "ACCOUNT_NUMBER_READ_WRITE_LENS", type = LensType.READ_WRITE),
                 @Lens(path = "currency.code", lensName = "ACCOUNT_CURRENCY_CODE_READ_LENS"),
                 @Lens(path = "currency.id", lensName = "ACCOUNT_CURRENCY_CODE_READ_WRITE_LENS", type = LensType.READ_WRITE)
         }
 )
-public class AccountWithSpecificFactoryName {
+class PackagePrivateAccount {
     private String id;
     private String accountNumber;
     private String bic;
