+++ lens4j-processor/src/test/resources/cases/package_private_class/PackagePrivateAccountLenses.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.package_private_class;
 
 import common.Currency;
 import dev.khbd.lens4j.core.Lenses;
