+++ lens4j-processor/src/test/resources/cases/single_property/Account.java
@@ -1,19 +1,23 @@
-package util.examples;
+package cases.single_property;
 
 import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
 import dev.khbd.lens4j.core.annotations.Lens;
 import dev.khbd.lens4j.core.annotations.LensType;
 
-@GenLenses(lenses = {
-        @Lens(path = "currencyFieldForTest.code"),
-        @Lens(path = "currencyFieldForTest.id", type = LensType.READ_WRITE)
-})
-public class AccountWithEmptyLensName {
+@GenLenses(
+        lenses = {
+                @Lens(path = "accountNumber"),
+                @Lens(path = "accountNumber", lensName = "ACCOUNT_NUMBER_READ_WRITE_LENS", type = LensType.READ_WRITE),
+                @Lens(path = "currency.code", lensName = "ACCOUNT_CURRENCY_CODE_READ_LENS"),
+                @Lens(path = "currency.id", lensName = "ACCOUNT_CURRENCY_CODE_READ_WRITE_LENS", type = LensType.READ_WRITE)
+        }
+)
+public class Account {
     private String id;
     private String accountNumber;
     private String bic;
-    private Currency currencyFieldForTest;
+    private Currency currency;
 
     public void setId(String id) {
         this.id = id;
@@ -39,11 +43,11 @@ public class AccountWithEmptyLensName {
         return bic;
     }
 
-    public void setCurrencyFieldForTest(Currency currencyFieldForTest) {
-        this.currencyFieldForTest = currencyFieldForTest;
+    public void setCurrency(Currency currency) {
+        this.currency = currency;
     }
 
-    public Currency getCurrencyFieldForTest() {
-        return currencyFieldForTest;
+    public Currency getCurrency() {
+        return currency;
     }
 }
