+++ lens4j-processor/src/test/resources/cases/specific_factory_name/AccountWithSpecificFactoryName.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.specific_factory_name;
 
 import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
@@ -6,14 +6,13 @@ import dev.khbd.lens4j.core.annotations.Lens;
 import dev.khbd.lens4j.core.annotations.LensType;
 
 @GenLenses(
+        factoryName = "SpecificFactoryName",
         lenses = {
-                @Lens(path = "accountNumber"),
-                @Lens(path = "accountNumber", lensName = "ACCOUNT_NUMBER_READ_WRITE_LENS", type = LensType.READ_WRITE),
                 @Lens(path = "currency.code", lensName = "ACCOUNT_CURRENCY_CODE_READ_LENS"),
                 @Lens(path = "currency.id", lensName = "ACCOUNT_CURRENCY_CODE_READ_WRITE_LENS", type = LensType.READ_WRITE)
         }
 )
-public class Account {
+public class AccountWithSpecificFactoryName {
     private String id;
     private String accountNumber;
     private String bic;
