+++ lens4j-processor/src/test/resources/cases/specific_factory_name/SpecificFactoryName.java
@@ -1,4 +1,4 @@
-package util.examples;
+package cases.specific_factory_name;
 
 import common.Currency;
 import dev.khbd.lens4j.core.Lenses;
@@ -10,10 +10,10 @@ import javax.annotation.processing.Generated;
 
 @Generated("dev.khbd.lens4j.processor.LensProcessor")
 public final class SpecificFactoryName {
-    public static final ReadLens<AccountWithDeCapitalizeSpecificFactoryName, String> ACCOUNT_CURRENCY_CODE_READ_LENS = Lenses.readLens(AccountWithDeCapitalizeSpecificFactoryName::getCurrency)
+    public static final ReadLens<AccountWithSpecificFactoryName, String> ACCOUNT_CURRENCY_CODE_READ_LENS = Lenses.readLens(AccountWithSpecificFactoryName::getCurrency)
             .andThen(Lenses.readLens(Currency::getCode));
 
-    public static final ReadWriteLens<AccountWithDeCapitalizeSpecificFactoryName, Integer> ACCOUNT_CURRENCY_CODE_READ_WRITE_LENS = Lenses.readLens(AccountWithDeCapitalizeSpecificFactoryName::getCurrency)
+    public static final ReadWriteLens<AccountWithSpecificFactoryName, Integer> ACCOUNT_CURRENCY_CODE_READ_WRITE_LENS = Lenses.readLens(AccountWithSpecificFactoryName::getCurrency)
             .andThen(Lenses.readWriteLens(Currency::getId, Currency::setId));
 
     private SpecificFactoryName() {
